+++ mars-expedition/src/settings/Settings.tsx
@@ -1,0 +1,3 @@
+export default function Settings() {
+    return <div>Setting page</div>
+}
