+++ mars-expedition/src/styles.css
@@ -1,0 +1,18 @@
+#app {
+    width: 100%;
+    height: 100%;
+    padding-top: 6rem;
+    
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    
+    position: relative;
+    overflow: hidden;
+    
+    background-color: #121212;
+}
+
+#game-container {
+    padding: 2rem;
+}
